requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
colorlog==6.7.0
croniter==1.3.8
discord-webhook==1.0.0
frozenlist==1.3.3
hikari==2.0.0.dev115
hikari-crescent==0.4.0
httpcore==0.16.3
httpx==0.23.3
numpy==1.24.1
pandas==1.5.2
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
types-croniter==1.0.11
typing_extensions==4.4.0
urllib3==1.24.3
yarl==1.8.2